Reorganized back-end class & method API
