+++ requirements.txt
@@ -1,2 +1,3 @@
 rapidwright==2022.1.1
+tqdm
 pytest
