--- test/rpd_test_report.txt
@@ -1,142 +1,0 @@
-======================================================================
-                             Bit Group 1
-======================================================================
-
-Significant Bits:
-------------------------------
-bit_0040241a_072_15 (0->1)
-	CLBLM_L_X72Y135 - SLICEL_X1.CLUT - INIT[00]
-	Resource Design Name: builder_csr_bankarray_interface8_bank_bus_dat_r[20]_i_5_TMR_2
-	INIT[00] bit altered for builder_csr_bankarray_interface8_bank_bus_dat_r[20]_i_5_TMR_2
-	Affected Resources:
-		builder_csr_bankarray_interface8_bank_bus_dat_r[20]_i_5_TMR_2
-
-	Vivado Tcl Commands:
-		select_objects [get_cells {builder_csr_bankarray_interface8_bank_bus_dat_r[20]_i_5_TMR_2}]
-
-Bits: 1
-Errors Found: 1 (100.0%)
-
-======================================================================
-                             Bit Group 2
-======================================================================
-
-Significant Bits:
-------------------------------
-bit_00402408_029_23 (1->0)
-	INT_L_X72Y114 - NR1BEG3 2-20 Routing Mux - Column Bit
-	Resource Design Name: INT_L_X72Y114/NR1BEG3
-	Opens created for net(s): builder_soclinux_adr_reg[2]_rep_n_0_TMR_1
-	Affected PIPs:
-		NE2END3->>NR1BEG3 (deactivated)
-	Affected Resources:
-		builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_1_TMR_1
-		builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_2_TMR_1
-		builder_csr_bankarray_interface11_bank_bus_dat_r[22]_i_2_TMR_1
-		builder_csr_bankarray_interface11_bank_bus_dat_r_reg_TMR_1[14]
-
-	Vivado Tcl Commands:
-		select_objects [get_pips {INT_L_X72Y114/INT_L.NE2END3->>NR1BEG3}]
-		select_objects [get_nets {builder_soclinux_adr_reg[2]_rep_n_0_TMR_1}]
-		select_objects [get_cells {builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_1_TMR_1 builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_2_TMR_1 builder_csr_bankarray_interface11_bank_bus_dat_r[22]_i_2_TMR_1 builder_csr_bankarray_interface11_bank_bus_dat_r_reg_TMR_1[14]}]
-
-Bits: 1
-Errors Found: 1 (100.0%)
-
-======================================================================
-                             Bit Group 3
-======================================================================
-
-Significant Bits:
-------------------------------
-bit_0040240e_029_23 (0->1)
-	INT_L_X72Y114 - NR1BEG3 2-20 Routing Mux - Row Bit
-	Resource Design Name: INT_L_X72Y114/NR1BEG3
-	Shorts formed between net(s): builder_soclinux_adr_reg[2]_rep_n_0_TMR_1 (initially connected), main_vfb_dma_sink_sink_last1_TMR_1[22]
-	Affected PIPs:
-		LOGIC_OUTS_L17->>NR1BEG3 (activated)
-	Affected Resources:
-		builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_1_TMR_1
-		builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_2_TMR_1
-		builder_csr_bankarray_interface11_bank_bus_dat_r[22]_i_2_TMR_1
-		builder_csr_bankarray_interface11_bank_bus_dat_r_reg_TMR_1[14]
-
-	Vivado Tcl Commands:
-		select_objects [get_pips {INT_L_X72Y114/INT_L.LOGIC_OUTS_L17->>NR1BEG3}]
-		select_objects [get_nets {builder_soclinux_adr_reg[2]_rep_n_0_TMR_1 main_vfb_dma_sink_sink_last1_TMR_1[22]}]
-		select_objects [get_cells {builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_1_TMR_1 builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_2_TMR_1 builder_csr_bankarray_interface11_bank_bus_dat_r[22]_i_2_TMR_1 builder_csr_bankarray_interface11_bank_bus_dat_r_reg_TMR_1[14]}]
-
-Bits: 1
-Errors Found: 1 (100.0%)
-
-======================================================================
-                             Bit Group 4
-======================================================================
-
-Significant Bits:
-------------------------------
-bit_0040240d_029_23 (0->1)
-	INT_L_X72Y114 - NR1BEG3 2-20 Routing Mux - Row Bit
-	Resource Design Name: INT_L_X72Y114/NR1BEG3
-	Shorts formed between net(s): Unconnected Wire(NL1BEG_N3), builder_soclinux_adr_reg[2]_rep_n_0_TMR_1 (initially connected)
-	Affected PIPs:
-		NL1BEG_N3->>NR1BEG3 (activated)
-	Affected Resources:
-		builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_1_TMR_1
-		builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_2_TMR_1
-		builder_csr_bankarray_interface11_bank_bus_dat_r[22]_i_2_TMR_1
-		builder_csr_bankarray_interface11_bank_bus_dat_r_reg_TMR_1[14]
-
-	Vivado Tcl Commands:
-		select_objects [get_pips {INT_L_X72Y114/INT_L.NL1BEG_N3->>NR1BEG3}]
-		select_objects [get_nets {builder_soclinux_adr_reg[2]_rep_n_0_TMR_1}]
-		select_objects [get_cells {builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_1_TMR_1 builder_csr_bankarray_interface11_bank_bus_dat_r[14]_i_2_TMR_1 builder_csr_bankarray_interface11_bank_bus_dat_r[22]_i_2_TMR_1 builder_csr_bankarray_interface11_bank_bus_dat_r_reg_TMR_1[14]}]
-
-Bits: 1
-Errors Found: 1 (100.0%)
-
-======================================================================
-                             Bit Group 5
-======================================================================
-
-Undefined Bits:
-------------------------------
-bit_0000002a_000_00
-	Potential Affected Resources:
-		No potential tiles found
-Bits: 1
-Errors Found: 0 (0.0%)
-
-======================================================================
-                             Bit Group 6
-======================================================================
-
-Errorless Bits:
-------------------------------
-bit_00423320_030_15: CLBLL_L_X102Y65 - SLICEL_X0.ALUT - INIT[00] - NA
-bit_00442c95_022_07: INT_R_X89Y11 - BYP_ALT0 5-24 Routing Mux - Row Bit - INT_R_X89Y11/BYP_ALT0
-
-Bits: 2
-Errors Found: 0 (0.0%)
-
-
-======================================================================
-              Design modelled: nexys_video_tmr_routed.dcp
-				Total time elapsed: 38.41 sec	(0 min)
-----------------------------------------------------------------------
-
-Bit Groups: 6
-Bit Groups w/ Errors: 4 (66.67%)
-
-Fault Bits: 7
-Routing Fault Bits: 4 (57.14%)
-CLB Fault Bits: 2 (28.57%)
-Unsupported Fault Bits: 0 (0.0%)
-Unknown Fault Bits: 1 (14.29%)
-Bits Driven High: 5 (71.43%)
-Bits Driven Low: 1 (14.29%)
-
-Found Errors: 4 (57.14%)
-PIP Open Errors: 1 (14.29%)
-PIP Short Errors: 2 (28.57%)
-CLB Altered Bit Errors: 1 (14.29%)
